BUILD normalize RSPEC closed metadata and guard skipped secrets
