+++ sonar-plugin/javascript-checks/src/main/resources/org/sonar/l10n/javascript/rules/javascript/S4790.json
@@ -1,6 +1,6 @@
 {
-  "title": "Using weak hashing algorithms is security-sensitive",
-  "type": "SECURITY_HOTSPOT",
+  "title": "Weak hashing algorithms should not be used",
+  "type": "VULNERABILITY",
   "code": {
     "impacts": {
       "SECURITY": "HIGH"
@@ -8,13 +8,15 @@
     "attribute": "TRUSTWORTHY"
   },
   "status": "ready",
-  "tags": [
-    "cwe"
-  ],
-  "extra": {
-    "replacementRules": [],
-    "legacyKeys": []
+  "quickfix": "unknown",
+  "remediation": {
+    "func": "Constant/Issue",
+    "constantCost": "30min"
   },
+  "tags": [
+    "cwe",
+    "former-hotspot"
+  ],
   "defaultSeverity": "Critical",
   "ruleSpecification": "RSPEC-4790",
   "sqKey": "S4790",
