+++ sonar-plugin/javascript-checks/src/main/resources/org/sonar/l10n/javascript/rules/javascript/Sonar_way_profile.json
@@ -1,1 +1,1 @@
-{"name":"Sonar way","ruleKeys":["S101","S1068","S107","S1077","S108","S1082","S1090","S1119","S1121","S1125","S1126","S1128","S1134","S1135","S1143","S1186","S1199","S1219","S1226","S125","S1264","S128","S1301","S1313","S1314","S1321","S1439","S1444","S1472","S1479","S1481","S1515","S1516","S1523","S1527","S1529","S1533","S1534","S1536","S1607","S1656","S1751","S1763","S1764","S1788","S1848","S1854","S1862","S1871","S1874","S1940","S1994","S2004","S2068","S2077","S2092","S2094","S2123","S2137","S2187","S2189","S2201","S2234","S2245","S2251","S2259","S2301","S2310","S2392","S2430","S2432","S2486","S2589","S2598","S2612","S2639","S2681","S2685","S2688","S2692","S2699","S2703","S2737","S2755","S2757","S2814","S2819","S2870","S2871","S2933","S2970","S2990","S2999","S3001","S3330","S3358","S3403","S3415","S3500","S3504","S3516","S3531","S3579","S3616","S3626","S3686","S3696","S3699","S3735","S3776","S3782","S3785","S3796","S3799","S3800","S3812","S3834","S3854","S3863","S3923","S3972","S3981","S3984","S4030","S4036","S4043","S4084","S4123","S4124","S4125","S4138","S4140","S4143","S4144","S4156","S4158","S4165","S4275","S4322","S4323","S4325","S4335","S4423","S4426","S4502","S4507","S4524","S4619","S4621","S4623","S4624","S4634","S4782","S4790","S4822","S4830","S5122","S5148","S5247","S5254","S5256","S5257","S5260","S5264","S5332","S5443","S5527","S5542","S5547","S5659","S5689","S5691","S5693","S5725","S5728","S5734","S5736","S5739","S5842","S5843","S5845","S5850","S5852","S5856","S5860","S5863","S5868","S5869","S5876","S5906","S5914","S5958","S5973","S6019","S6035","S6079","S6080","S6092","S6249","S6252","S6265","S6268","S6270","S6275","S6281","S6302","S6303","S6308","S6317","S6319","S6321","S6323","S6324","S6325","S6326","S6327","S6328","S6329","S6330","S6331","S6332","S6333","S6351","S6353","S6397","S6418","S6426","S6435","S6437","S6438","S6439","S6440","S6441","S6442","S6443","S6477","S6478","S6479","S6481","S6486","S6509","S6522","S6523","S6534","S6535","S6544","S6550","S6551","S6557","S6564","S6565","S6568","S6569","S6571","S6572","S6578","S6582","S6583","S6590","S6594","S6598","S6606","S6627","S6635","S6637","S6638","S6643","S6644","S6645","S6647","S6650","S6653","S6654","S6657","S6660","S6661","S6666","S6671","S6676","S6679","S6746","S6747","S6748","S6749","S6750","S6754","S6756","S6757","S6759","S6761","S6763","S6766","S6767","S6770","S6772","S6774","S6775","S6788","S6789","S6790","S6791","S6793","S6807","S6811","S6819","S6821","S6822","S6823","S6824","S6825","S6827","S6836","S6840","S6841","S6842","S6843","S6844","S6845","S6846","S6847","S6848","S6850","S6851","S6852","S6853","S6859","S6861","S6957","S6958","S6959","S7059","S7060","S7639","S7641","S7647","S7648","S7649","S7650","S7651","S7652","S7653","S7654","S7655","S7656","S7718","S7719","S7720","S7721","S7722","S7723","S7724","S7725","S7726","S7727","S7729","S7732","S7733","S7734","S7735","S7736","S7737","S7738","S7739","S7740","S7741","S7742","S7743","S7744","S7745","S7746","S7747","S7748","S7749","S7750","S7751","S7752","S7753","S7754","S7755","S7756","S7757","S7758","S7759","S7760","S7761","S7762","S7763","S7764","S7765","S7766","S7767","S7768","S7769","S7770","S7771","S7772","S7773","S7774","S7776","S7777","S7778","S7780","S7781","S7783","S7784","S7785","S7786","S7787","S7790","S8441","S8479","S8754","S878","S888","S905","S930"]}
+{"name":"Sonar way","ruleKeys":["S101","S1068","S107","S1077","S108","S1082","S1090","S1119","S1121","S1125","S1126","S1128","S1134","S1135","S1143","S1186","S1199","S1219","S1226","S125","S1264","S128","S1301","S1313","S1314","S1321","S1439","S1444","S1472","S1479","S1481","S1515","S1516","S1523","S1527","S1529","S1533","S1534","S1536","S1607","S1656","S1751","S1763","S1764","S1788","S1848","S1854","S1862","S1871","S1874","S1940","S1994","S2004","S2068","S2077","S2092","S2094","S2123","S2137","S2187","S2189","S2201","S2234","S2245","S2251","S2259","S2301","S2310","S2392","S2430","S2432","S2486","S2589","S2598","S2612","S2639","S2681","S2685","S2688","S2692","S2699","S2703","S2737","S2755","S2757","S2814","S2819","S2870","S2871","S2933","S2970","S2990","S2999","S3001","S3330","S3358","S3403","S3415","S3500","S3504","S3516","S3531","S3579","S3616","S3626","S3686","S3696","S3699","S3735","S3776","S3782","S3785","S3796","S3799","S3800","S3812","S3834","S3854","S3863","S3923","S3972","S3981","S3984","S4030","S4036","S4043","S4084","S4123","S4124","S4125","S4138","S4140","S4143","S4144","S4156","S4158","S4165","S4275","S4322","S4323","S4325","S4335","S4423","S4426","S4502","S4507","S4524","S4619","S4621","S4623","S4624","S4634","S4782","S4790","S4822","S4830","S5122","S5148","S5247","S5254","S5256","S5257","S5260","S5264","S5332","S5443","S5527","S5542","S5547","S5659","S5689","S5693","S5725","S5728","S5734","S5736","S5739","S5842","S5843","S5845","S5850","S5852","S5856","S5860","S5863","S5868","S5869","S5876","S5906","S5914","S5958","S5973","S6019","S6035","S6079","S6080","S6092","S6249","S6252","S6265","S6268","S6270","S6275","S6281","S6302","S6303","S6308","S6317","S6319","S6321","S6323","S6324","S6325","S6326","S6327","S6328","S6329","S6330","S6331","S6332","S6333","S6351","S6353","S6397","S6418","S6426","S6435","S6437","S6438","S6439","S6440","S6441","S6442","S6443","S6477","S6478","S6479","S6481","S6486","S6509","S6522","S6523","S6534","S6535","S6544","S6550","S6551","S6557","S6564","S6565","S6568","S6569","S6571","S6572","S6578","S6582","S6583","S6590","S6594","S6598","S6606","S6627","S6635","S6637","S6638","S6643","S6644","S6645","S6647","S6650","S6653","S6654","S6657","S6660","S6661","S6666","S6671","S6676","S6679","S6746","S6747","S6748","S6749","S6750","S6754","S6756","S6757","S6759","S6761","S6763","S6766","S6767","S6770","S6772","S6774","S6775","S6788","S6789","S6790","S6791","S6793","S6807","S6811","S6819","S6821","S6822","S6823","S6824","S6825","S6827","S6836","S6840","S6841","S6842","S6843","S6844","S6845","S6846","S6847","S6848","S6850","S6851","S6852","S6853","S6859","S6861","S6957","S6958","S6959","S7059","S7060","S7639","S7641","S7647","S7648","S7649","S7650","S7651","S7652","S7653","S7654","S7655","S7656","S7718","S7719","S7720","S7721","S7722","S7723","S7724","S7725","S7726","S7727","S7729","S7732","S7733","S7734","S7735","S7736","S7737","S7738","S7739","S7740","S7741","S7742","S7743","S7744","S7745","S7746","S7747","S7748","S7749","S7750","S7751","S7752","S7753","S7754","S7755","S7756","S7757","S7758","S7759","S7760","S7761","S7762","S7763","S7764","S7765","S7766","S7767","S7768","S7769","S7770","S7771","S7772","S7773","S7774","S7776","S7777","S7778","S7780","S7781","S7783","S7784","S7785","S7786","S7787","S7790","S8441","S8479","S8754","S878","S888","S905","S930"]}
